fix(lint): extract on-demand delegate code to keep server_compute.c under limit
The on-demand block pushed server_compute.c to 2115 lines, over the 2000-line
hard limit (CI line-check). Move delegate_spawn_ondemand / _drain /
_set_ceiling / _inflight into server_delegate_ondemand.c (declarations already
in server_compute_impl.h); server_compute.c is now 1998 lines. Wire the new
file into SERVER_SRCS (Makefile) and the CMake server source list. Tests that
link the referencing objects but not the new one get matching stubs:
test_server_compute.c stubs delegate_spawn_ondemand (capture, like its
compute_pool_submit stub); test_server_dispatch.c no-op-stubs
delegate_ondemand_set_ceiling / _drain.

Verified: line-check ok; unit-test-server-compute + unit-test-server-dispatch
pass; full server+kb link clean under -Werror.
