+++ CMakeLists.txt
@@ -793,6 +793,7 @@ set(SERVER_SRCS
     ${AIMEE_SRC_DIR}/server/server_coord_dispatcher.c
     ${AIMEE_SRC_DIR}/server/server_agent.c
     ${AIMEE_SRC_DIR}/server/server_compute.c
+    ${AIMEE_SRC_DIR}/server/server_delegate_ondemand.c
     ${AIMEE_SRC_DIR}/server/server_delegate_status.c
     ${AIMEE_SRC_DIR}/server/server_compute_async.c
     ${AIMEE_SRC_DIR}/server/server_jobs_aux.c
